+++ src/index.ts
@@ -27,13 +27,11 @@ export { decodeJwt, getEmbedOrigin } from "./jwt";
 export {
   confirmPaymentIntent,
   confirmSetupIntent,
-  hasNativeApplePaySession,
   sendConfirmationFailed,
   sendParentReadyMessage,
   updateAmount,
   updateAppearance,
   updateMerchantName,
-  updateNativeApplePaySession,
   validateForm,
 } from "./messaging";
 export type {